server/requirements.txt
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
config==0.5.1
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.8.0
pymongo==4.1.1
python-dotenv==0.20.0
six==1.16.0
toml==0.10.2
Werkzeug==2.1.2
